Timer Component 분리
Timer Component 분리
시, 분, 초 약간 추가 - 하나하나 설정해주는게 맞는지??
